requirements.txt
Django==1.11.23
paho-mqtt==1.5.0
celery==4.3.0
ipdb==0.12.2
